backend/requirements.txt
Flask
Flask-CORS
pymongo
PyJWT
bcrypt
PyMuPDF
python-dotenv
requests
certifi
